+++ packages/context-menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/context-menu
 
+## 1.0.19
+
+### Patch Changes
+
+- Add description props in select, textarea and textfield components
+- Updated dependencies
+  - @pikas-ui/icons@1.0.19
+  - @pikas-ui/loader@1.0.19
+  - @pikas-ui/menu@1.0.19
+  - @pikas-ui/styles@1.0.19
+
 ## 1.0.18
 
 ### Patch Changes