+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pikas-ui/styles
 
+## 1.0.19
+
+### Patch Changes
+
+- Add description props in select, textarea and textfield components
+
 ## 1.0.18
 
 ### Patch Changes