+++ packages/tooltip/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/tooltip
 
+## 1.0.19
+
+### Patch Changes
+
+- Add description props in select, textarea and textfield components
+- Updated dependencies
+  - @pikas-ui/styles@1.0.19
+
 ## 1.0.18
 
 ### Patch Changes